changed to csv format, renamed, and some changes
